move generated _down_
